25.02.2018

+++ 2. Programirane/9. Exam_Preparation/Task 08/Program.cs
@@ -5,15 +5,15 @@ using System.Linq;
 namespace Task_08
 {
     /// <summary>
-    /// Task 08. Anonymous Cache (Result: 70/100)
+    /// Task 08. Anonymous Cache (Result: 80/100)
     /// </summary>
     class Program
     {
         static void Main(string[] args)
         {
             // Речник и кеш
-            var data = new Dictionary<String, Dictionary<String, int>>();
-            var cache = new Dictionary<String, Dictionary<String, int>>();
+            var data = new Dictionary<String, Dictionary<String, long>>();
+            var cache = new Dictionary<String, Dictionary<String, long>>();
 
             // Обработка на входните данни
             string line = Console.ReadLine();
@@ -23,13 +23,13 @@ namespace Task_08
                 if (parts.Count() > 1)
                 {
                     String dataKey = parts[0];
-                    int dataSize = int.Parse(parts[1]);
+                    long dataSize = long.Parse(parts[1]);
                     String dataSet = parts[2];
 
                     // Добавяне към речника
                     if (data.ContainsKey(dataSet)) data[dataSet].Add(dataKey, dataSize);
                     // Добаввяне към кеша
-                    else if (!cache.ContainsKey(dataSet)) cache.Add(dataSet, new Dictionary<string, int>() { { dataKey, dataSize } });
+                    else if (!cache.ContainsKey(dataSet)) cache.Add(dataSet, new Dictionary<string, long>() { { dataKey, dataSize } });
                     else cache[dataSet].Add(dataKey, dataSize);
 
                 }
@@ -43,19 +43,19 @@ namespace Task_08
                         cache.Remove(dataSet);
                     }
                     // Ново множество
-                    else data.Add(dataSet, new Dictionary<string, int>());
+                    else data.Add(dataSet, new Dictionary<string, long>());
                 }
                 line = Console.ReadLine();
             }
 
             // Масимална сума (Labda Expression)
-            int max = data.Values.Max(x => x.Values.Sum());
+            long max = data.Values.Max(x => x.Values.Sum());
 
             // Печат
             foreach(var dataSet in data)
             {
                 // Сума на елементите в множеството (Labda Expression)
-                int sum = dataSet.Value.Sum(x => x.Value);
+                long sum = dataSet.Value.Sum(x => x.Value);
                 if (sum == max)
                 {
                     Console.WriteLine("Data Set: {0}, Total Size: {1}", dataSet.Key, sum);
